restructured statistics; created event_handlers module
