+++ src/event_handlers/mod.rs
@@ -1,0 +1,2 @@
+mod statistics;
+pub use statistics::StatsHandler;
